Significant infrastructure disturbance in support of coloring trails in a flexible way
Added the mapSeason property to the ConfigModel object to track the user
preference on how the seasons will be tracked (manual summer, manual winter, or
automatic based on date).  Also added the helper method isSummerMapSeason,
which determines if the trails should be based on the summer info.  For the
automatic mode, the determination is crudely determined as summer for dates
after the summer equinox and before the winter equinox.

Added the trailTypeEnabled property to the ConfigModel object, which is a
NSDictionary with keys being NSNumbers representing the trail type ID from the
database, and the values being NSNumbers interpreted as BOOL.

The TrailColorUtil singleton object was created to assist in determining the
color that a given trail type ID from the database should be rendered on the
map.  The singleton object can be retrieve by calling [TrailColorUtil
getTrailColorUtil], and the UIColor for a trail type can be retrieved with the
getTrailTypeColor method, which takes the trail type ID as an int.  As the
class exists now, the color for disabled trail types is a 75% gray with 50%
alpha.  All other trails are [UIColor redColor] for now, but this will be
extended to be dependent on the trail type IDs based on the heuristic
"difficulty" of the type (which will also have to factor in the season), and
the map type.

The MapViewController was shuffled around a bit to enable the querying of the
trail coordinates from the database iterated first by the trail types.

The SQLite database was temporarily swapped out for the stand-in DB again for
development purposes.