Work on the buttons in the pop-up POI info bubbles
Added a ConfigModel property sharingEnabled to track whether to put the share buttons in the POI info bubbles.

Modified the map POI info bubble code to add the "Share" button based on the
sharingEnabled configModel property, and the "More info" button based on
whether there is a URL in the database for that POI.

Significant reworking of the code that creates those buttons to do a better job
of sizing and aligning the buttons, so that they resize nicely relative to the
title text, and everything is optimally centered.  I'm going to rework this
code even just a bit more by making the buttons that are added more compactly
configurable, but I wanted to snapshot what I have first.

The database has been updated (manually) to contain the one URL that we have
for a POI.