Cleaned up the code that generates the buttons in the POI info bubble
Now the buttons are specified by an NSDictionary that specifies the predicate
that will determine if they are added to the info screen (an ObjC annonymous
block) and the NSString that selects the callback function that will be called
when the button is pressed.  In addition, there is an NSArray that specifies
the button names in the order that they should appear from left to right.

This will make updates of the code in the future for additional functionality
much quicker and less painful.