Updates to the GPS auto-tracking
Made a few updates to make the GPS auto-tracking code to make it respect the
user setting switch.  Pruned out some of the dead code, and organized the other
code better.  Found a typo in the Info.plist file that was the reason that the
permission had to be manually set, rather than being prompted, so now on
startup you are prompted by the system to allow location updates for the app.
The code to auto-disable the location tracking when the user pans away from the
GPS location, and then turn it back on when the MyLocation button is pressed is
not there yet, but that shouldn't be hard.  The MyLocation button causing the
didTapMyLocationButtonForMapView callback right now, though, so that'll have to
be figured out and fixed first.