Auto-size fixes for the map
Finally figured out how to fix the storyboard so that the MapView container
UIView auto-resizes to the available screen size!  The buttons on the map
screen are auto-placed by contraints properly now, as well.

As I was messing around with the placement of the buttons, I created an inset
border for the map so that the compass button wouldn't be placed under the
menubar at the top of the screen, and I inset it on the sides a slight amount
too, and now the buttons should line up with the compass when it shows up.

Also, when the map screen comes up, it hides the navigation control bar at the
top so that we get that space back for the map.