Added the Hybrid map type and cleaned up the constraints on the Settings page
The Settings page now allows for selection of the Hybrid map type in addition
to the others, and while I was updating the story board, I made sure that the
constraints were set up properly for everything to resize.  That included
changing the text of the "Automatically center map on GPS coordinates", which
was too long for the smaller iPhone screens.  Now it's shorter text with an
small text size annotation below it.