Added the first pass for filtering trail types and poi types
This changeset includes adding the poiTypeEnabled NSMutableDictionary to the
ConfigModel singleton, which is analogous to the trailTypeEnabled map.

Set up the FilterViewController to query the database for POI types and
populating the POI filter section of the UIListView.  Also added the code to
query the database for trail types appropriate for the season setting and
populating the trail filter section of the UIListView in coordination with the
TrailColorUtil singleton.

The TrailColorUtil singleton has been expanded to manage a number of things for
the FilterViewController.  The specifics of what it offers are probably best
left to code inspection, rather than detailing them all here in the commit
comments.

There are more changes to complete this functionality, including actually using
the poiTypeEnabled values when overlaying the POIs on the map, but this is a
stable snapshot of a detailed bunch of work, so I'm going to commit it here.