Updated to make GPS auto-tracking work the right way
Added two buttons (both temporarily without icons) for MyLocation and
DefaultCamera.  MyLocation is to replace the Google MyLocation button, since
Google's isn't working.  The DefaultCamera button takes the map to the default
camera that the app has when it first starts, which is set to zoom to fill the
screen with the entirety of the trail system.  This is currently set for an
iPhone6 screen, and needs to be update to make the zoom work for any device.

The GPS auto-tracking will now suspend when the user changes the map camera
with a gesture (such as panning or zooming manually), and the MyLocation button
will pan to the current GPS location.  If auto-tracking is turned on,
MyLocation will resume the tracking, and if not, the pan to the current GPS
location will be a single shot.

When panning to the GPS location for the first time, the zoom will be set
appropriately, but not later (so as to allow the user to select the zoom that
they want.)  This zoom should be re-worked also so that it only re-sets the
zoom level if the current GPS location is "significantly outside of the current
view" for some value of "significantly".