Did the main plumbing for putting the disc golf on the map
The polylines representing the disc golf holes are on the map, and the filter
page has been updated to allow turning them on and off.  Also the
TrailColorUtil has been updated to provide the UIColor to use for the path.

We still need some custom icons for the tees and baskets, and that should be
next.