Support for tapping the disc golf polylines to get the hole number
Incredible monkey business was needed to accomplish this.  Google simply
doesn't support this behavior, even though (once again) it seems that they do
in the API for Google Maps in the browser.  To get this to work, I needed to
register a guesture recognizer on the map, disable the gesture consuming
behavior of the map so that the recognizer would actually get called (that took
a while to figure out), cache the location of the tap in the handler, and then
when I get the callback for the tap on the polyline, I can put a phantom marker
(with a transparent 1x1 image as the icon) where the tap was and make it
selected, so tht my custom marker info window code can take over and draw the
info window.

Right now, this is only added for the disc golf holes, but all of the trails
have their titles loaded from the database, so I might publish an experiment
branch to see if everyone likes the ability to get pop-up info with the names
of trails on tapping them (it would be trivial to do now.)  Also, we can have
URLs for any of these things, and we'll automatically get "More info" links in
the pop-ups...